# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fleet Operations API
  version: 1.0.0
  description: 车队经营管理系统统一 API 契约。详细业务规则以项目需求基线文档为准。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: 手机号密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}}
...
  /users/{id}/reset-password:
    post:
      tags: [Auth]
      summary: 管理员重置密码
      x-permission: user:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /users/{id}/status:
    put:
      tags: [Auth]
      summary: 启用/停用账号
      x-permission: user:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
...
  /expenses/{id}/approval:
    get:
      tags: [Expenses]
      summary: 查看关联审批链
      x-permission: expense:approval:view
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
...
  /approvals:
    get:
      tags: [Approvals]
      summary: 审批列表
      description: scope=TODO/DONE/MINE 按当前登录用户服务端过滤；scope=ALL 仅具有 approval:history:view 的角色可查看全量。
      x-permission: approval:history:view / approval:process / approval:return / approval:create
      parameters:
        - {name: scope, in: query, required: false, schema: {$ref: '#/components/schemas/ApprovalListScope', default: ALL}}
        - {name: approvalType, in: query, required: false, schema: {type: string}}
        - {name: status, in: query, required: false, schema: {$ref: '#/components/schemas/ApprovalInstanceStatus'}}
        - {name: pageNo, in: query, required: false, schema: {type: integer, minimum: 1, default: 1}}
        - {name: pageSize, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/ApiResponse'}
                  - type: object
                    properties:
                      data: {$ref: '#/components/schemas/ApprovalPage'}
    post:
      tags: [Approvals]
      summary: 发起审批
      x-permission: approval:create
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalCreateRequest'}
      responses: {'200': {description: OK}}
  /approvals/{id}:
    get:
      tags: [Approvals]
      summary: 审批详情与历史
      description: 返回审批实例、全部 submission version、节点任务、实际审批人、动作、备注、时间和任务失效标记。
      x-permission: approval:history:view
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/ApiResponse'}
                  - type: object
                    properties:
                      data: {$ref: '#/components/schemas/ApprovalDetail'}
  /approvals/{id}/approve:
    post:
      tags: [Approvals]
      summary: 当前节点通过
      x-permission: approval:process
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalActionRequest'}
      responses: {'200': {description: OK}}
  /approvals/{id}/return-applicant:
    post:
      tags: [Approvals]
      summary: 打回申请人
      x-permission: approval:return
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalActionRequest'}
      responses: {'200': {description: OK}}
  /approvals/{id}/return-node:
    post:
      tags: [Approvals]
      summary: 打回指定前序节点
      x-permission: approval:return
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalActionRequest'}
      responses: {'200': {description: OK}}
  /approvals/{id}/resubmit:
    post:
      tags: [Approvals]
      summary: 修改后重新提交并生成新版本
      x-permission: approval:create
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalCreateRequest'}
      responses: {'200': {description: OK}}
  /approval-flows:
    get:
      tags: [Approvals]
      summary: 审批流程列表，可按类型和状态筛选
      x-permission: approval:flow:manage
      parameters:
        - {name: approvalType, in: query, required: false, schema: {type: string}}
        - {name: status, in: query, required: false, schema: {$ref: '#/components/schemas/ApprovalFlowStatus'}}
      responses: {'200': {description: OK}}
  /approval-flows/types/{approvalType}/versions:
    get:
      tags: [Approvals]
      summary: 查看某审批类型所有流程版本
      x-permission: approval:flow:manage
      parameters:
        - {name: approvalType, in: path, required: true, schema: {type: string}}
      responses: {'200': {description: OK}}
  /approval-flows/types/{approvalType}/active:
    get:
      tags: [Approvals]
      summary: 查看某审批类型当前 ACTIVE 版本及节点
      x-permission: approval:flow:manage
      parameters:
        - {name: approvalType, in: path, required: true, schema: {type: string}}
      responses: {'200': {description: OK}}
  /approval-flows/{id}/maintenance:
    post:
      tags: [Approvals]
      summary: 流程进入维护并禁止新发起
      x-permission: approval:flow:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /approval-flows/{id}:
    get:
      tags: [Approvals]
      summary: 查看流程版本详情、节点顺序和审批人
      x-permission: approval:flow:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
    put:
      tags: [Approvals]
      summary: 无在途审批后修改流程版本
      x-permission: approval:flow:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalFlowUpdateRequest'}
      responses: {'200': {description: OK}}
  /approval-flows/{id}/publish:
    post:
      tags: [Approvals]
      summary: 发布流程版本
      x-permission: approval:flow:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: integer, format: int64}}
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: {type: string, examples: ['0']}
        message: {type: string}
        data: {}
        requestId: {type: string}
    LoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone: {type: string}
        password: {type: string}
...
    AttachmentPurpose:
      type: string
      enum: [TIRE_OCR, WEIGHT_ADJUST, APPROVAL_APPLICATION, APPROVAL_ACTION, IMPORT_FILE]
    Attachment:
      type: object
      required: [id, ownerType, ownerId, purpose, storageKey, originalFilename, fileSize, fileHash, uploadedBy, uploadedAt]
      properties:
        id: {type: integer, format: int64}
        ownerType: {type: string}
        ownerId: {type: integer, format: int64}
        purpose: {$ref: '#/components/schemas/AttachmentPurpose'}
        originalFilename: {type: string}
        contentType: {type: [string, 'null']}
        fileSize: {type: integer, format: int64}
        fileHash: {type: string}
        storageKey: {type: string}
        uploadedBy: {type: integer, format: int64}
        uploadedAt: {type: string, format: date-time}
...
    ApprovalFlowStatus:
      type: string
      enum: [DRAFT, ACTIVE, MAINTENANCE]
    ApprovalInstanceStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, RETURNED_TO_APPLICANT]
    ApprovalTaskStatus:
      type: string
      enum: [PENDING, APPROVED, RETURNED, INVALIDATED]
    ApprovalListScope:
      type: string
      enum: [TODO, DONE, MINE, ALL]
    ApprovalCreateRequest:
      type: object
      required: [approvalType, businessType, businessId, businessSnapshot]
      properties:
        approvalType: {type: string}
        businessType: {type: string}
        businessId: {type: integer, format: int64}
        businessSnapshot:
          type: object
          additionalProperties: true
          description: 发起或重提时的申请快照；后端会把本次附件 ID 固化到 attachmentIds。
        attachmentIds:
          type: array
          items: {type: integer, format: int64}
          description: 已通过 ownerType=APPROVAL_UPLOAD、ownerId=当前用户ID、purpose=APPROVAL_APPLICATION 预上传的附件 ID。
    ApprovalActionRequest:
      type: object
      properties:
        comment: {type: [string, 'null']}
        targetNodeOrder: {type: [integer, 'null'], description: 打回指定节点时填写目标前序节点序号}
        attachmentIds:
          type: array
          items: {type: integer, format: int64}
          description: 审批动作前通过 ownerType=APPROVAL_UPLOAD、ownerId=当前用户ID、purpose=APPROVAL_ACTION 预上传的附件 ID；动作成功后绑定 approval_action.id。
    ApprovalSummary:
      type: object
      properties:
        id: {type: integer, format: int64}
        approvalNo: {type: string}
        approvalType: {type: string}
        businessType: {type: string}
        businessId: {type: integer, format: int64}
        applicantUserId: {type: integer, format: int64}
        status: {$ref: '#/components/schemas/ApprovalInstanceStatus'}
        currentNodeOrder: {type: [integer, 'null']}
        createdAt: {type: string, format: date-time}
    ApprovalPage:
      type: object
      properties:
        records:
          type: array
          items: {$ref: '#/components/schemas/ApprovalSummary'}
        total: {type: integer, format: int64}
        pageNo: {type: integer}
        pageSize: {type: integer}
    ApprovalFlowUpdateRequest:
      type: object
      required: [flowName, nodes]
      properties:
        flowName: {type: string}
        nodes:
          type: array
          minItems: 1
          items:
            type: object
            required: [nodeOrder, nodeName, approverUserId]
            properties:
              nodeOrder: {type: integer, minimum: 1}
              nodeName: {type: string}
              positionId: {type: [integer, 'null'], format: int64}
              approverUserId: {type: integer, format: int64}
              allowReturn: {type: boolean, default: true}
...
    ApprovalDetail:
      type: object
      properties:
        approval:
          type: object
          properties:
            id: {type: integer, format: int64}
            approvalNo: {type: string}
            approvalType: {type: string}
            flowId: {type: integer, format: int64}
            flowVersion: {type: integer}
            status: {$ref: '#/components/schemas/ApprovalInstanceStatus'}
            currentNodeOrder: {type: [integer, 'null']}
        submissions:
          type: array
          items:
            type: object
            properties:
              id: {type: integer, format: int64}
              versionNo: {type: integer}
              businessSnapshotJson: {type: string}
              submittedBy: {type: integer, format: int64}
              submittedAt: {type: string, format: date-time}
              attachments:
                type: array
                description: 当前 submission version 的申请人原始单据附件。
                items: {$ref: '#/components/schemas/Attachment'}
        tasks:
          type: array
          items:
            type: object
            properties:
              id: {type: integer, format: int64}
              submissionVersionId: {type: integer, format: int64}
              submissionVersionNo: {type: integer}
              nodeOrder: {type: integer}
              nodeName: {type: string}
              approverUserId: {type: integer, format: int64}
              status: {$ref: '#/components/schemas/ApprovalTaskStatus'}
              invalidated: {type: boolean}
              createdAt: {type: string, format: date-time}
              completedAt: {type: [string, 'null'], format: date-time}
              actions:
                type: array
                items:
                  type: object
                  properties:
                    id: {type: integer, format: int64}
                    actionType: {type: string, enum: [APPROVE, RETURN_APPLICANT, RETURN_NODE]}
                    operatorId: {type: integer, format: int64}
                    comment: {type: [string, 'null']}
                    targetNodeOrder: {type: [integer, 'null']}
                    operatedAt: {type: string, format: date-time}
                    attachments:
                      type: array
                      description: 当前 approval_action 的审批节点补充图片。
                      items: {$ref: '#/components/schemas/Attachment'}
...